♻️ Move functions from _utils.py to _models.py
The following functions are moved because they are only used in _models.py
* is_known_encoding
* parse_header_links
* parse_content_type_charset
* obfuscate_sensitive_headers

Related tests are also moved accordingly.
